pyproject.toml
[tool.poetry]
name = "gemcaen"
version = "0.2.0"
description = ""
authors = ["ivone"]

[[tool.poetry.packages]]
include = "gemcaen" 


[tool.poetry.dependencies]
python = ">=3.6.1 , <4"
pycaenhv = {git = "https://github.com/vasoto/pycaenhv.git"}
PyYAML = "^6.0"
setuptools = "^63.2.0"
tableformatter = "^0.1.6"
influxdb-client = "^1.31.0"
deepdiff = "4.3.0"
jsonpickle = "^2.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
